requirements.txt
matplotlib==3.8.2
mtcnn==0.1.1
numpy==1.23.1
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
pandas==1.5.0
Pillow==10.3.0
psycopg2==2.9.9
psycopg2_binary==2.9.9
reportlab==4.2.0
tkcalendar==1.6.1
torch==2.2.1
torchvision==0.17.1
